Check-in a fix for David Knopp.
Problem::
The proc2iso documentation and schema list the valid values of the ‘properties’ field in a <deviceprocessdata> entry to be ‘defaultset’, ‘setable’ or a combination of both.  proc2iso parses using strcmp() to check for valid values.  This means that only the first of possibly two settings is being parsed.  Rather, the parsing should be handled like the ‘trigger_methods’ field, where an strstr() loop parses all valid substrings.

Solution:
Parse the ‘properties’ field in a manner similar to ‘trigger_methods’, using substring search instead of string comparison.