Integrate two way for SW and pool version integration as #defines: a) Linux Makefile includes in same directory as the Makefile a file named versions.mk - when this file exists b) the project config header includes now also a header named versions.h, that is automatically created    in the same directory as the config header. An already existing version.h header is untouched.
The variant a) works only for unix makefiles, while b) works for all platforms.
But with a), the version numbers can be defined outside the source tree in the individual built directory.
This can be wanted - but can trigger also inconsistencies, when the same code is compiled in different build
directories for different target setups.

The versions.mk file must constist of a line like:
VERSION_DEFINES = -DPOOL_VERSION="1.0" -DFIRMWARE_VERSION="1.5"

The content of the auto-generated version.h header (in the same directory as the project config header) is as:
#ifndef POOL_VERSION
        #define POOL_VERSION=1.0
#endif
#ifndef FIRMWARE_VERSION
        #define FIRMWARE_VERSION=1.0
#endif

==>>
this way any version specification in version.mk overrides all settings in the version.h header.