the IsoAgLib had until now several classes with names which didn't match a strict camel case pattern - example ISOMonitor_c which got now renamed to IsoMonitor_c
We are providing API class name backward compatibility for some time, so that you don't have to change your application code immediately - but you should do this soon. This is achieved by typedefs like 
"typedef IsoMonitor_c ISOMonitor_c;".

The following class names are provided in old naming with typedef:
+ typedef [i]CanIo_c [i]CANIO_c;
+ typedef [i]CanPkg_c [i]CANPkg_c;
+ typedef [i]CanPkgExt_c [i]CANPkgExt_c;
+ typedef [i]CanCustomer_c [i]CANCustomer_c;
+ typedef [i]IsoMonitor_c [i]ISOMonitor_c;
+ typedef [i]IsoItem_c [i]ISOItem_c;
+ typedef [i]IsoName_c [i]ISOName_c;
+ typedef [i]IsoTerminal_c [i]ISOTerminal_c;
+ typedef [i]EepromIo_c [i]EEPROMIO_c;
+ typedef IsoFilter_s ISOFilter_s;
+ typedef IsoSystemPkg_c ISOSystemPkg_c;
+ typedef ElementDdi_s ElementDDI_s;
+ iIsoName_c::iISONameUnspecified(), iIsoName_c::iISONameInitialProcessData()
  in addition to the renamed API functions
  iIsoName_c::iIsoNameUnspecified(), iIsoName_c::iIsoNameInitialProcessData()

The following renames are NOT providing a compatibility typedef, as they shouldn't be used by normal apps - please contact us, if this is different in your case, so that we can add backward compatibility typedefs:
+ "s/ISOFilterBox_c/IsoFilterBox_c/g"
+ "s/ISOFilterManager_c/IsoFilterManager_c/g"
+ "s/ISORequestPGN_c/IsoRequestPgn_c/g"
+ "s/ISORequestPGNHandler_c/IsoRequestPgnHandler_c/g"
+ "s/ISOTerminalPkg_c/IsoTerminalPkg_c/g"