Merge over from Trunk: r4392,r4396 [Update2standard] [TC] [Adapted message-parsing from DIS/FDIS to IS]
David Knopp:
- Correct the handling of the Task Status and Working-set Maintenance messages:
When determining the message type, only the low nibble of byte 0 should be
checked, rather than the whole byte. The upper nibble is the low four bits
of the element number. The element number in these messages was changed in
the part 10 standard from DIS->IS. Devices conforming to the IS will no
longer work using IsoAgLib as implemented.
It's a simple change and will not affect TC clients conforming to previous
DIS and FDIS.

Worked so far in Trunk.
No review needed.